+++ src/compas_viewer/components/sceneform.py
@@ -1,11 +1,9 @@
 from typing import Callable
 from typing import Optional
 
-from PySide6.QtGui import QColor
+from PySide6.QtCore import Qt
 from PySide6.QtWidgets import QTreeWidget
 from PySide6.QtWidgets import QTreeWidgetItem
-
-from compas.scene import Scene
 
 
 class Sceneform(QTreeWidget):
@@ -15,128 +13,115 @@ class Sceneform(QTreeWidget):
     Parameters
     ----------
     scene : :class:`compas.scene.Scene`
-        The tree to be displayed. An typical example is the scene
-        object tree: :attr:`compas_viewer.viewer.Viewer._tree`.
-    columns : dict[str, callable]
+        The scene to be displayed.
+    columns : list[dict]
         A dictionary of column names and their corresponding attributes.
-        Example: ``{"Name": (lambda o: o.name), "Object": (lambda o: o)}``
-    column_editable : list, optional
-        A list of booleans indicating whether the corresponding column is editable.
-        Defaults to ``[False]``.
+        Example: {"Name": lambda o: o.name, "Object": lambda o: o}
+    column_editable : list[bool], optional
+        A list of booleans indicating whether the corresponding column is editable. Defaults to [False].
     show_headers : bool, optional
-        Show the header of the tree.
-        Defaults to ``True``.
-    stretch : int, optional
-        Stretch factor of the tree in the grid layout.
-        Defaults to ``2``.
-    backgrounds : dict[str, callable], optional
-        A dictionary of column names and their corresponding color.
-        Example: ``{"Object-Color": (lambda o: o.surfacecolor)}``
+        Show the header of the tree. Defaults to True.
+    callback : Callable, optional
+        Callback function to execute when an item is clicked or selected.
 
     Attributes
     ----------
-    tree : :class:`compas.datastructures.Tree`
-        The tree to be displayed.
-
-    See Also
-    --------
-    :class:`compas.datastructures.Tree`
-    :class:`compas.datastructures.tree.TreeNode`
-    :class:`compas_viewer.layout.SidedockLayout`
-
-    References
-    ----------
-    :PySide6:`PySide6/QtWidgets/QTreeWidget`
-
-    Examples
-    --------
-    .. code-block:: python
-
-        from compas_viewer import Viewer
-
-        viewer = Viewer()
-
-        for i in range(10):
-            for j in range(10):
-                sp = viewer.scene.add(Sphere(0.1, Frame([i, j, 0], [1, 0, 0], [0, 1, 0])), name=f"Sphere_{i}_{j}")
-
-        viewer.layout.sidedock.add_element(Treeform(viewer._tree, {"Name": (lambda o: o.object.name), "Object": (lambda o: o.object)}))
-
-        viewer.show()
-
+    scene : :class:`compas.scene.Scene`
+        The scene to be displayed.
+    columns : list[dict]
+        A dictionary of column names and their corresponding function.
+    checkbox_columns : dict[int, dict[str, Callable]]
+        A dictionary of column indices and their corresponding attributes.
     """
 
     def __init__(
         self,
-        scene: Scene,
-        columns: dict[str, Callable],
-        column_editable: list[bool] = [False],
+        columns: list[dict],
+        column_editable: Optional[list[bool]] = None,
         show_headers: bool = True,
-        stretch: int = 2,
-        backgrounds: Optional[dict[str, Callable]] = None,
         callback: Optional[Callable] = None,
     ):
         super().__init__()
         self.columns = columns
-        self.column_editable = column_editable + [False] * (len(columns) - len(column_editable))
+        self.checkbox_columns: dict[int, str] = {}
+        self.column_editable = (column_editable or [False]) + [False] * (len(columns) - len(column_editable or [False]))
         self.setColumnCount(len(columns))
-        self.setHeaderLabels(list(self.columns.keys()))
+        self.setHeaderLabels(col["title"] for col in self.columns)
         self.setHeaderHidden(not show_headers)
-        self.stretch = stretch
-        self._backgrounds = backgrounds
 
-        self.scene = scene
         self.callback = callback
-        self.itemClicked.connect(self.on_item_clickded)
+
+        self.itemClicked.connect(self.on_item_clicked)
         self.itemSelectionChanged.connect(self.on_item_selection_changed)
 
     @property
-    def scene(self) -> Scene:
-        return self._scene
+    def viewer(self):
+        from compas_viewer import Viewer
 
-    @scene.setter
-    def scene(self, scene: Scene):
+        return Viewer()
+
+    @property
+    def scene(self):
+        return self.viewer.scene
+
+    def update(self):
         self.clear()
-        for node in scene.traverse("breadthfirst"):
+        self.checkbox_columns = {}
+
+        for node in self.scene.traverse("breadthfirst"):
             if node.is_root:
                 continue
 
-            strings = [str(c(node)) for _, c in self.columns.items()]
+            strings = []
 
-            if node.parent.is_root:  # type: ignore
-                node.attributes["widget"] = QTreeWidgetItem(self, strings)  # type: ignore
-            else:
-                node.attributes["widget"] = QTreeWidgetItem(
-                    node.parent.attributes["widget"],
-                    strings,  # type: ignore
-                )
+            for i, column in enumerate(self.columns):
+                type = column.get("type", None)
+                if type == "checkbox":
+                    action = column.get("action")
+                    checked = column.get("checked")
+                    if not action or not checked:
+                        raise ValueError("Both action and checked must be provided for checkbox")
+                    self.checkbox_columns[i] = {"action": action, "checked": checked}
+                    strings.append("")
+                elif type == "label":
+                    text = column.get("text")
+                    if not text:
+                        raise ValueError("Text must be provided for label")
+                    strings.append(text(node))
 
-            node.attributes["widget"].node = node
-            node.attributes["widget"].setSelected(node.is_selected)
+            parent_widget = self if node.parent.is_root else node.parent.attributes["widget"]
+            widget = QTreeWidgetItem(parent_widget, strings)
+            widget.node = node
+            widget.setSelected(node.is_selected)
+            widget.setFlags(widget.flags() | Qt.ItemIsUserCheckable | Qt.ItemIsSelectable | Qt.ItemIsEnabled)
 
-            if self._backgrounds:
-                for col, background in self._backgrounds.items():
-                    node.attributes["widget"].setBackground(list(self.columns.keys()).index(col), QColor(*background(node).rgb255))
+            for col, col_data in self.checkbox_columns.items():
+                widget.setCheckState(col, Qt.Checked if col_data["checked"](node) else Qt.Unchecked)
 
-        self._scene = scene
+            node.attributes["widget"] = widget
 
-    def update(self):
-        from compas_viewer import Viewer
+        self.adjust_column_widths()
 
-        self.scene = Viewer().scene
+    def on_item_clicked(self, item, column):
+        if column in self.checkbox_columns:
+            check = self.checkbox_columns[column]["action"]
+            check(item.node, item.checkState(column) == Qt.Checked)
 
-    def on_item_clickded(self):
-        selected_nodes = [item.node for item in self.selectedItems()]
-        for node in self.scene.objects:
-            node.is_selected = node in selected_nodes
-            if self.callback and node.is_selected:
-                self.callback(node)
+        if self.selectedItems():
+            selected_nodes = {item.node for item in self.selectedItems()}
+            for node in self.scene.objects:
+                node.is_selected = node in selected_nodes
+                if self.callback and node.is_selected:
+                    self.callback(node)
 
-        from compas_viewer import Viewer
-
-        Viewer().renderer.update()
+        self.viewer.renderer.update()
 
     def on_item_selection_changed(self):
         for item in self.selectedItems():
             if self.callback:
                 self.callback(item.node)
+
+    def adjust_column_widths(self):
+        for i in range(self.columnCount()):
+            if i in self.checkbox_columns:
+                self.setColumnWidth(i, 50)